+++ tests/integration/tools/test_config_backup.py
@@ -28,13 +28,13 @@ class TestHappyPathUnipiConfigBackup:
     )
     def test_unipi_config_backup(self, config_loader: ConfigLoader, caplog: LogCaptureFixture) -> None:
         """Test for config backup."""
-        backup_path: Path = config_loader.temp_path / "backup"
-        backup_path.mkdir()
+        backup_dir: Path = config_loader.persistent_tmp_dir / "backup"
+        backup_dir.mkdir()
 
-        main([backup_path.as_posix(), "-c", config_loader.temp.as_posix()])
+        main([backup_dir.as_posix(), "-c", config_loader.temp.as_posix()])
 
         logs: List[str] = [record.getMessage() for record in caplog.records]
-        files: List[Path] = list(backup_path.glob("*.tar.gz"))
+        files: List[Path] = list(backup_dir.glob("*.tar.gz"))
 
         assert len(logs) == 1
         assert len(files) == 1
@@ -53,7 +53,7 @@ class TestUnhappyPathUnipiConfigBackup:
     def test_cli_output_directory_not_exists(self, config_loader: ConfigLoader, caplog: LogCaptureFixture) -> None:
         """Test for missing output directory."""
         with pytest.raises(SystemExit) as error:
-            main([(config_loader.temp_path / "not_exists").as_posix(), "-c", config_loader.temp.as_posix()])
+            main([(config_loader.persistent_tmp_dir / "not_exists").as_posix(), "-c", config_loader.temp.as_posix()])
 
         logs: List[str] = [record.getMessage() for record in caplog.records]
 
@@ -71,7 +71,7 @@ class TestUnhappyPathUnipiConfigBackup:
     def test_cli_output_is_not_a_directory(self, config_loader: ConfigLoader, caplog: LogCaptureFixture) -> None:
         """Test for output is not a directory."""
         with pytest.raises(SystemExit) as error:
-            main([config_loader.config_file_path.as_posix(), "-c", config_loader.temp.as_posix()])
+            main([config_loader.config_file.as_posix(), "-c", config_loader.temp.as_posix()])
 
         logs: List[str] = [record.getMessage() for record in caplog.records]
 
@@ -95,11 +95,11 @@ class TestUnhappyPathUnipiConfigBackup:
         mock_tarfile_open = mocker.patch.object(tarfile, "open")
         mock_tarfile_open.return_value.__enter__.return_value.add = mock_tar_add
 
-        backup_path: Path = config_loader.temp_path / "backup"
-        backup_path.mkdir()
+        backup_dir: Path = config_loader.persistent_tmp_dir / "backup"
+        backup_dir.mkdir()
 
         with pytest.raises(SystemExit) as error:
-            main([backup_path.as_posix(), "-c", config_loader.temp.as_posix()])
+            main([backup_dir.as_posix(), "-c", config_loader.temp.as_posix()])
 
         logs: List[str] = [record.getMessage() for record in caplog.records]