+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Changed temporary directory from `/tmp/unipi` to the persistent `/var/tmp/unipi` directory.
 - Bump pymodbus to version 3.3.2
 - Bump ayncio-mqtt to version 1.0.0 (the package name changed to aiomqtt)
+
+## Fixed
+
+- Repair the OPKG package for the Unipi Control OS. Added the missing control files to the package. 
 
 ## [2.0.0] - 2023-06-23